move checks for files existing on commit into internal module
